added profile image
