+++ pages/edit-profile.js
@@ -9,8 +9,11 @@ import { connect } from 'react-redux';
 import Types from '../src/actions/index';
 import Input from '../src/components/Input';
 import Textarea from '../src/components/Textarea';
+import config from '../src/config';
 import { editProfileAction } from '../src/actions/user';
 
+
+const configOptions = config[process.env.NODE_ENV || 'development'];
 
 class EditProfile extends React.Component {
   state = {
@@ -32,12 +35,13 @@ class EditProfile extends React.Component {
     if (!user.data || isEmpty(user.data)) {
       Router.push('/login');
     }
-    const { first_name, last_name, email, bio } = user.data;
+    const { first_name, last_name, email, bio, image } = user.data;
     this.setState({
       firstName: first_name,
       lastName: last_name,
       email,
       bio,
+      image,
     });
   }
 
@@ -54,15 +58,46 @@ class EditProfile extends React.Component {
       name: user.data.name,
     });
 
-    Router.push('/profile');
+    Router.push('/my-profile');
+  }
+
+  handleEditProfileImage = async (e) => {
+    const { user } = this.props;
+    const files = Array.from(e.target.files)
+
+    if (files.length > 1) {
+      const msg = 'Only 1 image can be uploaded at a time'
+      alert(msg)
+    }
+
+    const image = files[0]
+
+    const formData = new FormData();
+    formData.append('image', image)
+
+    const config = {
+      headers: {
+        'content-type': 'multipart/form-data'
+      }
+    }
+
+    const url = `${configOptions.hostname}/api/users/${user.data.id}/profilepicture`;
+
+    const profilepicResp = await axios.post(url, formData, config);
   }
 
   render() {
-    const { firstName, lastName, email, bio } = this.state;
+    const { firstName, lastName, email, bio, image } = this.state;
 
     return (
       <div className="padded-section thin-section">
         <h1>Edit Profile</h1>
+        <Input
+          label="Upload Profile Image"
+          type='file'
+          id='single'
+          accept="image/png, image/jpeg"
+          onChange={(e) => this.handleEditProfileImage(e)}/> 
         <Input
           label="First name"
           value={firstName}
