+++ pages/my-profile.js
@@ -57,33 +57,6 @@ class MyProfile extends React.Component {
     })
   }
 
-  handleEditProfileImage = async (e) => {
-    const { user } = this.props;
-    const files = Array.from(e.target.files)
-
-    if (files.length > 1) {
-      const msg = 'Only 1 image can be uploaded at a time'
-      alert(msg)
-    }
-
-    const image = files[0]
-
-    const formData = new FormData();
-    formData.append('file', image)
-
-    const config = {
-      headers: {
-        'content-type': 'multipart/form-data'
-      }
-    }
-
-    const url = `${configOptions.hostname}/api/users/${user.data.id}/profilepicture`;
-
-    const profilepicResp = await axios.post(url, formData, config);
-
-    console.log(profilepicResp)
-  }
-
   componentDidMount = async () => {
     const { user, followings } = this.props;
     if (!user.data || isEmpty(user.data)) {
@@ -126,7 +99,7 @@ class MyProfile extends React.Component {
 
     let userInfo = this.props.user.data;
 
-    const { first_name, last_name, email } = userInfo;
+    const { first_name, last_name, email, image } = userInfo;
 
     let { bio } = userInfo;
 
@@ -138,11 +111,7 @@ class MyProfile extends React.Component {
         <div className="padded-section">
           {
           <div className="profile-image-wrapper">
-            <img className="profile-image" src="static/blankprofile.png"/>
-            <div className="edit-profile-hover clickable" >
-            <label for="single">Edit</label>
-              <input type='file' id='single' accept="image/png, image/jpeg" onChange={(e) => this.handleEditProfileImage(e)}/> 
-            </div>
+            <img className="profile-image" src={ image ? `${configOptions.hostname}${image}` : "static/blankprofile.png"}/>
           </div>
           }
           {
