+++ src/components/Input.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styles from 'styles/base.scss';
 
-const Input = ({ value, type, placeholder, onChange, label, error, className, buttonClick, buttonLabel, style }) => (
+const Input = ({ value, type, placeholder, onChange, label, accept, error, className, buttonClick, buttonLabel, style }) => (
 
   <div className="form-input-wrapper" style={style}>
     <p className="form-label">{label}</p>
@@ -10,6 +10,7 @@ const Input = ({ value, type, placeholder, onChange, label, error, className, bu
       <input
         value={value}
         type={type}
+        accept={accept}
         placeholder={placeholder}
         className={className || 'form-input'}
         onChange={onChange}
@@ -40,6 +41,7 @@ Input.propTypes = {
   buttonLabel: PropTypes.string,
   label: PropTypes.string,
   type: PropTypes.string,
+  accept: PropTypes.string,
   className: PropTypes.string,
   value: PropTypes.string,
   placeholder: PropTypes.string,
